requirements.txt
groq
python-dotenv
pytest
streamlit
watchdog
